resolve issues from feedback, update shared package
