+++ packages/front-end/components/ActivityList.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 import { AuditInterface } from "back-end/types/audit";
 import Link from "next/link";
+import { date, datetime } from "shared";
 import useApi from "../hooks/useApi";
-import { date, datetime } from "../services/dates";
 import LoadingOverlay from "./LoadingOverlay";
 import Avatar from "./Avatar/Avatar";
 //import { phaseSummary } from "../services/utils";
