+++ packages/front-end/components/EventWebHooks/EventWebHookDetail/EventWebHookDetail.tsx
@@ -7,11 +7,11 @@ import classNames from "classnames";
 import { useRouter } from "next/router";
 import Link from "next/link";
 import { HiOutlineClipboard, HiOutlineClipboardCheck } from "react-icons/hi";
+import { datetime } from "shared";
 import DeleteButton from "@/components/DeleteButton/DeleteButton";
 import { useAuth } from "@/services/auth";
 import Badge from "@/components/Badge";
 import { EventWebHookEditParams, useIconForState } from "../utils";
-import { datetime } from "../../../services/dates";
 import { useCopyToClipboard } from "../../../hooks/useCopyToClipboard";
 import { SimpleTooltip } from "../../SimpleTooltip/SimpleTooltip";
 import useApi from "../../../hooks/useApi";
