+++ packages/front-end/components/EventWebHooks/EventWebHookLogs/EventWebHookLogItem/EventWebHookLogItem.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react";
 import { NotificationEvent } from "back-end/src/events/notification-events";
 import { EventWebHookLogInterface } from "back-end/types/event-webhook-log";
 import classNames from "classnames";
-import { datetime } from "@/services/dates";
+import { datetime } from "shared";
 import { useIconForState } from "../../utils";
 
 type EventWebHookLogItemProps = {
