+++ packages/front-end/components/Experiment/AnalysisSettingsBar.tsx
@@ -9,9 +9,12 @@ import {
 import { StatsEngine } from "back-end/types/stats";
 import { FaInfoCircle } from "react-icons/fa";
 import { OrganizationSettings } from "back-end/types/organization";
-import { DEFAULT_SEQUENTIAL_TESTING_TUNING_PARAMETER } from "shared";
+import {
+  DEFAULT_SEQUENTIAL_TESTING_TUNING_PARAMETER,
+  ago,
+  datetime,
+} from "shared";
 import { useAuth } from "@/services/auth";
-import { ago, datetime } from "@/services/dates";
 import { useDefinitions } from "@/services/DefinitionsContext";
 import usePermissions from "@/hooks/usePermissions";
 import Toggle from "@/components/Forms/Toggle";
