+++ packages/front-end/components/Experiment/NewFeatureExperiments.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { ExperimentRule, FeatureInterface } from "back-end/types/feature";
 import { ExperimentInterfaceStringDates } from "back-end/types/experiment";
+import { ago } from "shared";
 import useApi from "@/hooks/useApi";
 import { useDefinitions } from "@/services/DefinitionsContext";
-import { ago } from "@/services/dates";
 import usePermissions from "@/hooks/usePermissions";
 import NewExperimentForm from "./NewExperimentForm";
