+++ packages/front-end/components/Experiment/Results.tsx
@@ -3,9 +3,8 @@ import React, { FC, useEffect } from "react";
 import dynamic from "next/dynamic";
 import { StatsEngine } from "back-end/types/stats";
 import { MetricRegressionAdjustmentStatus } from "back-end/types/report";
-import { getValidDate } from "shared";
+import { getValidDate, ago } from "shared";
 import { useDefinitions } from "@/services/DefinitionsContext";
-import { ago } from "@/services/dates";
 import usePermissions from "@/hooks/usePermissions";
 import useOrgSettings from "@/hooks/useOrgSettings";
 import { useAuth } from "@/services/auth";
