+++ packages/front-end/components/Features/RevisionDropdown.tsx
@@ -3,9 +3,9 @@ import { FeatureRevisionInterface } from "back-end/types/feature-revision";
 import clsx from "clsx";
 import { useMemo, useState } from "react";
 import { MdRestore } from "react-icons/md";
+import { ago, datetime } from "shared";
 import usePermissions from "@/hooks/usePermissions";
 import { useAuth } from "@/services/auth";
-import { ago, datetime } from "@/services/dates";
 import Dropdown from "../Dropdown/Dropdown";
 import Modal from "../Modal";
 import Pagination from "../Pagination";
