+++ packages/front-end/components/HistoryTable.tsx
@@ -4,8 +4,8 @@ import Link from "next/link";
 import ReactDiffViewer, { DiffMethod } from "react-diff-viewer";
 import { BsArrowRepeat } from "react-icons/bs";
 import { FaAngleDown, FaAngleUp } from "react-icons/fa";
+import { ago, datetime } from "shared";
 import { useDefinitions } from "../services/DefinitionsContext";
-import { ago, datetime } from "../services/dates";
 import useApi from "../hooks/useApi";
 import Button from "./Button";
 import Code from "./SyntaxHighlighting/Code";
