+++ packages/front-end/components/Layout/AccountPlanNotices.tsx
@@ -1,9 +1,9 @@
 import { useRouter } from "next/router";
 import { useState } from "react";
 import { FaExclamationTriangle } from "react-icons/fa";
+import { daysLeft } from "shared";
 import usePermissions from "@/hooks/usePermissions";
 import useStripeSubscription from "@/hooks/useStripeSubscription";
-import { daysLeft } from "@/services/dates";
 import { isCloud } from "@/services/env";
 import { useUser } from "@/services/UserContext";
 import UpgradeModal from "../Settings/UpgradeModal";
