+++ packages/front-end/components/Share/ShareModal.tsx
@@ -15,8 +15,7 @@ import { GrDrag } from "react-icons/gr";
 import { FaCheck, FaRegTrashAlt } from "react-icons/fa";
 import { FiAlertTriangle } from "react-icons/fi";
 import { HexColorPicker } from "react-colorful";
-import { getValidDate } from "shared";
-import { ago, datetime, date } from "@/services/dates";
+import { getValidDate, ago, datetime, date } from "shared";
 import { useAuth } from "@/services/auth";
 import { useUser } from "@/services/UserContext";
 import { useSearch } from "@/services/search";
