+++ packages/front-end/pages/dimensions/index.tsx
@@ -3,8 +3,8 @@ import { FaPencilAlt } from "react-icons/fa";
 import { DimensionInterface } from "back-end/types/dimension";
 import clsx from "clsx";
 import Link from "next/link";
+import { ago } from "shared";
 import LoadingOverlay from "@/components/LoadingOverlay";
-import { ago } from "@/services/dates";
 import Button from "@/components/Button";
 import DimensionForm from "@/components/Dimensions/DimensionForm";
 import { useDefinitions } from "@/services/DefinitionsContext";
