+++ packages/front-end/pages/features/index.tsx
@@ -4,8 +4,8 @@ import { useRouter } from "next/router";
 import { useFeature } from "@growthbook/growthbook-react";
 import { FaExclamationTriangle } from "react-icons/fa";
 import { FeatureInterface } from "back-end/types/feature";
+import { ago, datetime } from "shared";
 import LoadingOverlay from "@/components/LoadingOverlay";
-import { ago, datetime } from "@/services/dates";
 import { GBAddCircle } from "@/components/Icons";
 import FeatureModal from "@/components/Features/FeatureModal";
 import ValueDisplay from "@/components/Features/ValueDisplay";
