+++ packages/front-end/pages/idea/[iid].tsx
@@ -11,6 +11,7 @@ import Link from "next/link";
 import { ImpactEstimateInterface } from "back-end/types/impact-estimate";
 import { ExperimentInterfaceStringDates } from "back-end/types/experiment";
 import { useForm } from "react-hook-form";
+import { date } from "shared";
 import useApi from "@/hooks/useApi";
 import LoadingOverlay from "@/components/LoadingOverlay";
 import { useAuth } from "@/services/auth";
@@ -18,7 +19,6 @@ import DeleteButton from "@/components/DeleteButton/DeleteButton";
 import DiscussionThread from "@/components/DiscussionThread";
 import useSwitchOrg from "@/services/useSwitchOrg";
 import ImpactModal from "@/components/Ideas/ImpactModal";
-import { date } from "@/services/dates";
 import NewExperimentForm from "@/components/Experiment/NewExperimentForm";
 import ViewQueryButton from "@/components/Metrics/ViewQueryButton";
 import ImpactProjections from "@/components/Ideas/ImpactProjections";
