+++ packages/front-end/pages/ideas.tsx
@@ -3,9 +3,9 @@ import Link from "next/link";
 import { IdeaInterface } from "back-end/types/idea";
 import { FaPlus, FaRegCheckSquare, FaRegSquare } from "react-icons/fa";
 import clsx from "clsx";
+import { date } from "shared";
 import useApi from "../hooks/useApi";
 import LoadingOverlay from "../components/LoadingOverlay";
-import { date } from "../services/dates";
 import IdeaForm from "../components/Ideas/IdeaForm";
 import { useSearch } from "../services/search";
 import { useDefinitions } from "../services/DefinitionsContext";
