+++ packages/front-end/pages/presentations.tsx
@@ -2,12 +2,12 @@ import React, { useState } from "react";
 import Link from "next/link";
 import { PresentationInterface } from "back-end/types/presentation";
 import { FaPlus } from "react-icons/fa";
+import { date } from "shared";
 import useApi from "../hooks/useApi";
 import LoadingOverlay from "../components/LoadingOverlay";
 import ShareModal from "../components/Share/ShareModal";
 import ConfirmModal from "../components/ConfirmModal";
 import { useAuth } from "../services/auth";
-import { date } from "../services/dates";
 import Modal from "../components/Modal";
 import CopyToClipboard from "../components/CopyToClipboard";
 import { useUser } from "../services/UserContext";
