+++ packages/front-end/pages/reports.tsx
@@ -3,8 +3,8 @@ import { useRouter } from "next/router";
 import Link from "next/link";
 import { ReportInterface } from "back-end/types/report";
 import { ExperimentInterface } from "back-end/types/experiment";
+import { datetime, ago } from "shared";
 import LoadingOverlay from "../components/LoadingOverlay";
-import { datetime, ago } from "../services/dates";
 import { useAddComputedFields, useSearch } from "../services/search";
 import Tooltip from "../components/Tooltip/Tooltip";
 import useApi from "../hooks/useApi";
