+++ packages/front-end/pages/segments/index.tsx
@@ -5,8 +5,8 @@ import { IdeaInterface } from "back-end/types/idea";
 import { MetricInterface } from "back-end/types/metric";
 import Link from "next/link";
 import clsx from "clsx";
+import { ago } from "shared";
 import LoadingOverlay from "@/components/LoadingOverlay";
-import { ago } from "@/services/dates";
 import Button from "@/components/Button";
 import SegmentForm from "@/components/Segments/SegmentForm";
 import { useDefinitions } from "@/services/DefinitionsContext";
