--- packages/front-end/services/dates.ts
@@ -1,30 +1,0 @@
-import format from "date-fns/format";
-import formatDistance from "date-fns/formatDistance";
-import differenceInDays from "date-fns/differenceInDays";
-import { addMonths } from "date-fns";
-import { getValidDate } from "shared";
-
-export function date(date: string | Date): string {
-  if (!date) return "";
-  return format(getValidDate(date), "PP");
-}
-export function datetime(date: string | Date): string {
-  if (!date) return "";
-  return format(getValidDate(date), "PPp");
-}
-export function ago(date: string | Date): string {
-  if (!date) return "";
-  return formatDistance(getValidDate(date), new Date(), { addSuffix: true });
-}
-export function daysLeft(date: string | Date): number {
-  return differenceInDays(getValidDate(date), new Date());
-}
-export function subtractMonths(date: string | Date, num: number): Date {
-  return addMonths(getValidDate(date), -1 * num);
-}
-export function monthYear(date: string | Date): string {
-  return format(getValidDate(date), "MMM yyy");
-}
-export function daysBetween(start: string | Date, end: string | Date): number {
-  return differenceInDays(getValidDate(end), getValidDate(start));
-}
